pyproject.toml
[tool.poetry]
name = "project-oop"
version = "0.1.0"
description = ""
authors = ["User <tooxa85@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"
isort = "^5.13.2"
black = "^24.10.0"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
line_length = 119

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]